requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-core
langchain-groq
pydantic
python-dotenv
sqlalchemy
# Optional: Use Alembic if you want to manage DB migrations in future
# alembic
